requirements.txt
pandas
dash
mysql-connector-python